# docs/swagger.yaml
basePath: /
definitions:
...
  kinds.CreateUserRequest:
    properties:
      department:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      organization:
        type: string
      password:
        type: string
      title:
        type: string
      user_uid:
        type: string
      username:
        type: string
    type: object
  kinds.DatastoreServiceRequest:
    properties:
      irods_path:
        type: string
      irods_user:
        type: string
    type: object
  kinds.DatastoreUserRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
...
  kinds.EmailRequest:
    properties:
      bcc:
        items:
          type: string
        type: array
      from_email:
        type: string
      html_body:
        type: string
      subject:
        type: string
      text_body:
        type: string
      to:
        items:
          type: string
        type: array
    type: object
  kinds.GenericResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
...
  kinds.PasswordChangeRequest:
    properties:
      password:
        type: string
    type: object
...
  kinds.UserExistsResponse:
    properties:
      exists:
        type: boolean
      username:
        type: string
    type: object
...
  kinds.UserResponse:
    properties:
      user:
        type: string
    type: object
  kinds.ValidateResponse:
    properties:
      valid:
        type: boolean
    type: object
  kinds.ValidationError:
    properties:
      loc:
        example:
        - body
        - username
        items:
          type: string
        type: array
      msg:
        example: Field required
        type: string
      type:
        example: missing
        type: string
    type: object
  kinds.ValidationErrorResponse:
    properties:
      detail:
        items:
          $ref: '#/definitions/kinds.ValidationError'
        type: array
    type: object
info:
  contact: {}
  description: API for managing CyVerse users, LDAP, DataStore, and email notifications.
  title: Portal Conductor API
  version: "1.0"
paths:
...
  /datastore/users:
    post:
      consumes:
      - application/json
      description: Create user in iRODS / DataStore only (or reset password if user
        already exists).
      parameters:
      - description: Datastore User details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/kinds.DatastoreUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/kinds.UserResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/kinds.ValidationErrorResponse'
        "500":
          description: Failed to create datastore user
          schema:
            $ref: '#/definitions/kinds.GenericResponse'
      security:
      - BasicAuth: []
      summary: Create datastore user
  /datastore/users/{username}/exists:
    get:
      description: Check whether a username exists in the DataStore.
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/kinds.UserExistsResponse'
        "500":
          description: Failed to check user existence in data store
          schema:
            $ref: '#/definitions/kinds.GenericResponse'
      security:
      - BasicAuth: []
      summary: Check user datastore existence
  /datastore/users/{username}/services:
    post:
      consumes:
      - application/json
      description: Register DataStore services with the iRODS path and username.
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      - description: Datastore Service Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/kinds.DatastoreServiceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/kinds.GenericResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/kinds.ValidationErrorResponse'
        "500":
          description: Failed to register datastore service
          schema:
            $ref: '#/definitions/kinds.GenericResponse'
      security:
      - BasicAuth: []
      summary: Register datastore service
  /emails/send:
    post:
      consumes:
      - application/json
      description: Send an email using SMTP.
      parameters:
      - description: Email details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/kinds.EmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/kinds.GenericResponse'
        "400":
          description: Either text_body or html_body must be provided
          schema:
            $ref: '#/definitions/kinds.GenericResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/kinds.ValidationErrorResponse'
        "500":
          description: Failed to send email
          schema:
            $ref: '#/definitions/kinds.GenericResponse'
      security:
      - BasicAuth: []
      summary: Send email
...
  /users:
    post:
      consumes:
      - application/json
      description: Create a complete user account in both LDAP (with default groups)
        and DataStore.
      parameters:
      - description: User Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/kinds.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/kinds.UserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/kinds.GenericResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/kinds.ValidationErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/kinds.GenericResponse'
      security:
      - BasicAuth: []
      summary: Create user
  /users/{username}:
    delete:
      description: Synchronously remove the user from the datastore and LDAP.
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/kinds.UserResponse'
      security:
      - BasicAuth: []
      summary: Delete user synchronously
  /users/{username}/password:
    post:
      consumes:
      - application/json
      description: Update the user's password in both LDAP and the datastore.
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      - description: New password details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/kinds.PasswordChangeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/kinds.UserResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/kinds.ValidationErrorResponse'
      security:
      - BasicAuth: []
      summary: Change password
  /users/{username}/validate:
    post:
      consumes:
      - application/json
      description: Validate user credentials against LDAP.
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      - description: Password details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/kinds.PasswordChangeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/kinds.ValidateResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/kinds.ValidationErrorResponse'
      security:
      - BasicAuth: []
      summary: Validate credentials
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"